pre-fetch data working solution

+++ server/server.js
@@ -1,16 +1,17 @@
 // react imports
 import { StaticRouter } from 'react-router-dom';
+import { Provider } from 'react-redux';
 
 const React = require('react');
-
 const express = require('express');
 const path = require('path');
 const fs = require('fs');
 const { exec } = require('child_process');
 require('isomorphic-fetch');
+const bodyParser = require('body-parser');
+// const hasPreloadRan = require('../lib/api/hasPreloadRan.js');
 
 const app = express();
-
 // User app directory is received from arguments
 const appParentDirectory = process.argv[2];
 const ribbitConfig = require(path.join(appParentDirectory, '/ribbit.config.js'));
@@ -18,9 +19,7 @@ const ribbitRoutes = require(path.join(
   appParentDirectory,
   `${ribbitConfig.appRoot}/ribbit.routes.json`
 ));
-
 const appFile = `${appParentDirectory}/${ribbitConfig.appRoot}/${ribbitConfig.app}`;
-
 // Helper functions imports
 const buildRoutesCliCommand = require('./helpers/buildRoutesCliCommand');
 const sendFetches = require('./helpers/sendFetches');
@@ -53,23 +52,50 @@ const routesCliCommand = buildRoutesCliCommand(
   ribbitConfig.appRoot
 );
 
+const preloadArray = [];
+
+app.use(bodyParser.json());
+
+app.get(['/preload-push', '/preload-pop'], (req, res) => {
+  const arrayCommand = req.url.substring(req.url.lastIndexOf('-') + 1);
+  if (arrayCommand === 'push') {
+    preloadArray.push(1);
+    res.end();
+  } else if (arrayCommand === 'pop') {
+    preloadArray.pop();
+    res.end();
+  }
+});
+
 app.get(
   routeArray,
   (req, res, next) => {
     const CompiledApp = require(`../dist/App.js`).default;
-    const context = {};
+
+    //  user exports their store from wherever they created it
+    // user must give the path to their store file
+    const { store } = require(`../dist/App.js`);
+
+    const context = { data: {}, head: [], req };
     let componentRoute = req.url;
 
+    // pull state out of store
+    const preLoadedState = store.getState();
+
     const jsx = (
-      <StaticRouter context={context} location={componentRoute}>
-        <CompiledApp />
-      </StaticRouter>
+      // wrap static router in redux Provider in order to user redux state
+      <Provider store={store}>
+        <StaticRouter context={context} location={componentRoute}>
+          <CompiledApp />
+        </StaticRouter>
+      </Provider>
     );
 
     if (componentRoute === '/') componentRoute = routesCliCommand.homeComponent;
 
     res.locals = {
       ...res.locals,
+      preLoadedState,
       appParentDirectory,
       componentRoute,
       jsx,
@@ -80,18 +106,13 @@ app.get(
   htmlTemplate,
   writeFile
 );
+
 app.use(express.static(ribbitConfig.bundleRoot));
 
-// Create a new child process, that executes the passed in 'cli command'
-// Child starts webpack and copies components over to the Ribbit directory
-
 const webpackChild = exec(`${routesCliCommand.command}`, () => {
-  // start server in callback (after webpack finishes running)
-  app.listen(4000, () => {
-    console.log('Listening on port 4000');
-    // Send fetch request to all routes
-    const fetchArray = sendFetches(ribbitRoutes, 4000);
-
+  app.listen(5000, () => {
+    console.log('Listening on port 5000');
+    const fetchArray = sendFetches(ribbitRoutes, 5000);
     Promise.all(fetchArray)
       .then(arrayOfRoutes => {
         const ribbitManifest = arrayOfRoutes.reduce((acc, curr) => {
@@ -105,10 +126,19 @@ const webpackChild = exec(`${routesCliCommand.command}`, () => {
           `${appParentDirectory}/ribbit.manifest.json`,
           JSON.stringify(ribbitManifest)
         );
-        unlinkUserDeps(ribbitConfig, appParentDirectory);
-        process.kill(process.pid, 'SIGINT');
+
+        function killServer() {
+          if (preloadArray.length === 0) {
+            console.log('KILL THE SERVER GENTS!!!!');
+            process.kill(process.pid, 'SIGINT');
+          } else {
+            console.log('NOT READY TO KILL SERVER');
+            setTimeout(killServer, 500);
+          }
+        }
+        killServer();
       })
-      .catch();
+      .catch(err => console.log(err));
   });
 });
